Now we can edit endpoints
