+++ pages/api/new.js
@@ -30,7 +30,7 @@ export default async function handler(req, res) {
   if (existsSync(ENDPOINT_DIRECTORY)) {
     try {
       console.log(`${ENDPOINT_DIRECTORY}/${endpointPath}.json`);
-      writeFileSync(`${ENDPOINT_DIRECTORY}/${endpointPath}.json`, JSON.stringify(endpointData));
+      writeFileSync(`${ENDPOINT_DIRECTORY}/${endpointPath}.json`, JSON.stringify(endpointData, null, 2));
       responseStatus = 200;
       responseData = { message: 'Endpoint created succesfully!'};
     } catch (err) {
